requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
gunicorn==21.2.0